feat: centralize print slot logic
