+++ index.html
@@ -591,6 +591,7 @@
     <!-- ▸▸▸ SCRIPTS ------------------------------------------------------ -->
     <script type="module" src="js/modelViewerTouchFix.js"></script>
     <script type="module" src="js/wizard.js" defer></script>
+    <script type="module" src="js/print-slots.js" defer></script>
     <script type="module" src="js/index.js" defer></script>
     <script type="module" src="js/subredditLanding.js" defer></script>
     <script type="module" src="js/exitDiscount.js" defer></script>
