+++ js/payment.js
@@ -1,4 +1,9 @@
 import { track } from "./analytics.js";
+import {
+  computeSlotsByTime,
+  adjustedSlots,
+  recordSlotPurchase,
+} from "./print-slots.js";
 
 (() => {
   try {
@@ -267,77 +272,6 @@ function ensureModelViewerLoaded() {
   });
 }
 
-function getCycleKey() {
-  const now = new Date();
-  const dateFmt = new Intl.DateTimeFormat("en-US", {
-    timeZone: TZ,
-    year: "numeric",
-    month: "2-digit",
-    day: "2-digit",
-  });
-  const hourFmt = new Intl.DateTimeFormat("en-US", {
-    timeZone: TZ,
-    hour: "numeric",
-    hour12: false,
-  });
-  const dateStr = dateFmt.format(now);
-  const hour = parseInt(hourFmt.format(now), 10);
-  if (hour < 1) {
-    const prev = new Date(now.getTime() - 24 * 60 * 60 * 1000);
-    return dateFmt.format(prev);
-  }
-  return dateStr;
-}
-
-function resetPurchaseCount() {
-  const key = getCycleKey();
-  if (localStorage.getItem("slotCycle") !== key) {
-    localStorage.setItem("slotCycle", key);
-    localStorage.setItem("slotPurchases", "0");
-  }
-}
-
-function getPurchaseCount() {
-  resetPurchaseCount();
-  const n = parseInt(localStorage.getItem("slotPurchases"), 10);
-  return Number.isInteger(n) && n > 0 ? n : 0;
-}
-
-function recordPurchase() {
-  resetPurchaseCount();
-  const n = getPurchaseCount();
-  localStorage.setItem("slotPurchases", String(n + 1));
-  const sessionId = localStorage.getItem("adSessionId");
-  const subreddit = localStorage.getItem("adSubreddit");
-  if (sessionId && subreddit) {
-    track("checkout", { sessionId, subreddit, step: "complete" }).catch(
-      () => {},
-    );
-  }
-}
-
-function adjustedSlots(base) {
-  const n = getPurchaseCount();
-  return Math.max(0, base - n);
-}
-
-function computeSlotsByTime() {
-  const dtf = new Intl.DateTimeFormat("en-US", {
-    timeZone: TZ,
-    hour12: false,
-    hour: "numeric",
-  });
-  const hour = parseInt(dtf.format(new Date()), 10);
-  if (hour >= 1 && hour < 4) return 9;
-  if (hour >= 4 && hour < 7) return 8;
-  if (hour >= 7 && hour < 10) return 7;
-  if (hour >= 10 && hour < 13) return 6;
-  if (hour >= 13 && hour < 16) return 5;
-  if (hour >= 16 && hour < 19) return 4;
-  if (hour >= 19 && hour < 22) return 3;
-  if (hour >= 22 && hour < 24) return 2;
-  return 1;
-}
 
 function computeColorSlotsByTime() {
   const dtf = new Intl.DateTimeFormat("en-US", {
@@ -350,6 +284,17 @@ function computeColorSlotsByTime() {
   if (hour >= 12) return 4;
   if (hour >= 9) return 5;
   return 6;
+}
+
+function recordPurchase() {
+  recordSlotPurchase();
+  const sessionId = localStorage.getItem("adSessionId");
+  const subreddit = localStorage.getItem("adSubreddit");
+  if (sessionId && subreddit) {
+    track("checkout", { sessionId, subreddit, step: "complete" }).catch(
+      () => {},
+    );
+  }
 }
 
 function qs(name) {
