+++ payment.html
@@ -743,6 +743,7 @@
     </script>
     <script src="js/modelViewerTouchFix.js" defer></script>
     <script type="module" src="js/wizard.js" defer></script>
+    <script type="module" src="js/print-slots.js" defer></script>
     <script type="module" src="js/payment.js" defer></script>
     <script type="module" src="js/exitDiscount.js" defer></script>
     <script>
